+++ admin/api/key/delete.php
@@ -7,18 +7,16 @@ $data = (object)array_map(function ($v) {
 }, $_POST);
 
 global $arr, $database;
-if (checkDatabase($arr)) {
-    if (!empty($data->voteKey)) {
-        $voteKeys = $data->voteKey;
-        $statement = $database->getConnection()->prepare("DELETE FROM `voting_keys` WHERE `VoteKey` IN (" . getQMarks($data->voteKey) . ")");
-        bindArray($statement, $data->voteKey);
-        if ($statement->execute()) {
-            http_response_code(200);
-            $arr = array("message" => (is_array($data->voteKey) ? "Keys" : "Key ($data->voteKey)") . " wurde(n) entfernt.");
-        } else {
-            http_response_code(503);
-            $arr = array("message" => (is_array($data->voteKey) ? "Keys" : "Key ($data->voteKey)") . " konnte(n) nicht entfernt werden.");
-        }
+if (!empty($data->voteKey) && checkDatabase($arr)) {
+    $voteKeys = $data->voteKey;
+    $statement = $database->getConnection()->prepare("DELETE FROM `voting_keys` WHERE `VoteKey` IN (" . getQMarks($data->voteKey) . ")");
+    bindArray($statement, $data->voteKey);
+    if ($statement->execute()) {
+        http_response_code(200);
+        $arr = array("message" => (is_array($data->voteKey) ? "Keys" : "Key ($data->voteKey)") . " wurde(n) entfernt.");
+    } else {
+        http_response_code(503);
+        $arr = array("message" => (is_array($data->voteKey) ? "Keys" : "Key ($data->voteKey)") . " konnte(n) nicht entfernt werden.");
     }
 }
